Make process context abstract; move implementation to default.
This opens the possibility for other process contexts existing.
