+++ Libraries/libutil/Sources/ProcessContext.cpp
@@ -8,124 +8,20 @@
  */
 
 #include <libutil/ProcessContext.h>
-#include <libutil/FSUtil.h>
 
 #include <sstream>
 #include <unordered_set>
-#include <cstring>
-#include <cassert>
-
-#include <pwd.h>
-#include <grp.h>
-#include <unistd.h>
-#include <sys/select.h>
-#include <errno.h>
-
-#if defined(__APPLE__)
-#include <mach-o/dyld.h>
-#elif defined(__linux__)
-#include <linux/limits.h>
-#if __GLIBC__ >= 2 && __GLIBC_MINOR__ >= 16
-#include <sys/auxv.h>
-#endif
-#endif
-
-extern "C" char **environ;
 
 using libutil::ProcessContext;
-using libutil::FSUtil;
 
-std::string ProcessContext::
-currentDirectory() const
+ProcessContext::
+ProcessContext()
 {
-    char path[PATH_MAX + 1];
-    if (::getcwd(path, sizeof(path)) == nullptr) {
-        path[0] = '\0';
-    }
-    return path;
 }
 
-#if defined(__linux__)
-static char initialWorkingDirectory[PATH_MAX] = { 0 };
-__attribute__((constructor))
-static void InitializeInitialWorkingDirectory()
+ProcessContext::
+~ProcessContext()
 {
-    if (getcwd(initialWorkingDirectory, sizeof(initialWorkingDirectory)) == NULL) {
-        abort();
-    }
-}
-
-#if !(__GLIBC__ >= 2 && __GLIBC_MINOR__ >= 16)
-static char initialExecutablePath[PATH_MAX] = { 0 };
-__attribute__((constructor))
-static void InitialExecutablePathInitialize(int argc, char **argv)
-{
-    strncpy(initialExecutablePath, argv[0], sizeof(initialExecutablePath));
-}
-#endif
-#endif
-
-std::string ProcessContext::
-executablePath() const
-{
-#if defined(__APPLE__)
-    uint32_t size = 0;
-    if (_NSGetExecutablePath(NULL, &size) != -1) {
-        abort();
-    }
-
-    std::string buffer;
-    buffer.resize(size);
-    if (_NSGetExecutablePath(&buffer[0], &size) != 0) {
-        abort();
-    }
-
-    return FSUtil::NormalizePath(buffer);
-#elif defined(__linux__)
-#if __GLIBC__ >= 2 && __GLIBC_MINOR__ >= 16
-    char const *path = reinterpret_cast<char const *>(getauxval(AT_EXECFN));
-    if (path == NULL) {
-        assert(false);
-        abort();
-    }
-#elif defined(__GLIBC__)
-    char const *path = reinterpret_cast<char const *>(initialExecutablePath);
-#else
-#error Unsupported platform.
-#endif
-
-    std::string absolutePath = FSUtil::ResolveRelativePath(std::string(path), std::string(initialWorkingDirectory));
-    return FSUtil::NormalizePath(absolutePath);
-#else
-#error Unsupported platform.
-#endif
-}
-
-ext::optional<std::string> ProcessContext::
-environmentVariable(std::string const &variable) const
-{
-    if (char *value = getenv(variable.c_str())) {
-        return std::string(value);
-    } else {
-        return ext::nullopt;
-    }
-}
-
-std::unordered_map<std::string, std::string> ProcessContext::
-environmentVariables() const
-{
-    std::unordered_map<std::string, std::string> environment;
-
-    for (char **current = environ; *current; current++) {
-        std::string variable = *current;
-        std::string::size_type offset = variable.find('=');
-
-        std::string name = variable.substr(0, offset);
-        std::string value = variable.substr(offset + 1);
-        environment.insert({ name, value });
-    }
-
-    return environment;
 }
 
 std::vector<std::string> ProcessContext::
@@ -151,70 +47,18 @@ executableSearchPaths() const
     return paths;
 }
 
-std::string ProcessContext::
-userName() const
-{
-    std::string result;
+#include <libutil/DefaultProcessContext.h>
 
-    if (struct passwd const *pw = ::getpwuid(::getuid())) {
-        if (pw->pw_name != nullptr) {
-            result = std::string(pw->pw_name);
-        }
-    }
-
-    if (result.empty()) {
-        std::ostringstream os;
-        os << ::getuid();
-        result = os.str();
-    }
-
-    ::endpwent();
-
-    return result;
-}
-
-std::string ProcessContext::
-groupName() const
-{
-    std::string result;
-
-    if (struct group const *gr = ::getgrgid(::getgid())) {
-        if (gr->gr_name != nullptr) {
-            result = gr->gr_name;
-        }
-    }
-
-    if (result.empty()) {
-        std::ostringstream os;
-        os << ::getgid();
-        result = os.str();
-    }
-
-    ::endgrent();
-
-    return result;
-}
-
-int32_t ProcessContext::
-userID() const
-{
-    return ::getuid();
-}
-
-int32_t ProcessContext::
-groupID() const
-{
-    return ::getgid();
-}
+using libutil::DefaultProcessContext;
 
 ProcessContext const *ProcessContext::
 GetDefault()
 {
-    static ProcessContext *sysUtil = nullptr;
-    if (sysUtil == nullptr) {
-        sysUtil = new ProcessContext();
+    static DefaultProcessContext *defaultProcessContext = nullptr;
+    if (defaultProcessContext == nullptr) {
+        defaultProcessContext = new DefaultProcessContext();
     }
 
-    return sysUtil;
+    return defaultProcessContext;
 }
